Add ActionViews, rename transaction_view module to view
