+++ transaction/src/view/action_view/swap_claim_view.rs
